backend/requirements.txt
# FastAPI backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI and ML libraries
transformers==4.36.0
torch==2.1.0
langchain==0.0.350
langgraph==0.0.40
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# Database
pymongo==4.6.0
motor==3.3.2

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Logging and monitoring
structlog==23.2.0

# AWS integration
boto3==1.34.0
botocore==1.34.0

# Medical data processing
fhir-resources==7.0.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
